[FLINK-8421] [core] Allow InstantiationUtil.deserializeObject() to use FailureTolerantObjectInputStream
This makes it possible to use the FailureTolerantObjectInputStream in
InstantiationUtil.deserializeObject. This behaviour is controlled via a
isFailureTolerant flag argument. All previous calls to the method
remains to have identical behaviour (the ClassloaderObjectInputStream).
